Create db (#109)
* Added a proxy for backend routes

* Migrated local database to remote

* Fixed DB query and redirect page issues

* removed hard coded database credentials

* update account working

* Delete .DS_Store

* Delete .DS_Store

* delete lowercase login

* delete lowercase login

* readded Login

* Delete .DS_Store

* updated to latest dependencies in backend package.json

* added event page back end

* removed pug files

* removed pug dependency from json

* update to use database lib connection

* cleaned up URL for account apis

* updated routing urls for account
