+++ packages/backend/src/app.js
@@ -11,8 +11,7 @@ import indexRouter from './routes/index';
 import usersRouter from './routes/users';
 // import ticketsRouter from './routes/tickets';
 // import homeRouter from './routes/home';
-import createAccountRouter from './routes/create-account';
-import loginRouter from './routes/login';
+import accountRouter from './routes/account';
 import searchRouter from './routes/search';
 
 const port = process.env.PORT || 2000;
@@ -42,8 +41,7 @@ process.on('unhandledRejection', err => {
   app.use('/users', usersRouter);
   // app.use('/tickets', ticketsRouter);
   // app.use('/home', homeRouter);
-  app.use('/create-account', createAccountRouter);
-  app.use('/login', loginRouter);
+  app.use('/account', accountRouter);
   app.use('/search', searchRouter);
 
   // wait until the app starts
